backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# HTTP
requests==2.32.3
httpx==0.27.2
aiohttp==3.10.5

# ML / Data Science
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
xgboost==2.1.1

# Google Gemini AI
google-generativeai==0.8.3

# Database (SQLite for dev, can swap to Postgres)
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Scheduling / Background tasks
apscheduler==3.10.4

# CORS
starlette==0.38.6

# Logging
loguru==0.7.2
